requirements.txt
langgraph>=0.4
pydantic>=2.7
python-dotenv>=1.0
openai>=1.40
pillow>=10.3